+++ src/clj_foundationdb/utils.clj
@@ -11,7 +11,6 @@
   [value]
   (.getInt (ByteBuffer/wrap value)))
 
-
 (defn bytes-to-str
   [bytes]
   (apply str (map char bytes)))