+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [Unreleased]
 
+- Add clear-key, get-range, get-range-startswith, get-all, clear-range
 - Add method to set a value, get a value, get a range of keys, clear database